Move remaining analytics packages to solr.analytics
