+++ base/src/main/java/uk/ac/ebi/atlas/species/SpeciesInferrer.java
@@ -3,7 +3,7 @@ package uk.ac.ebi.atlas.species;
 import com.google.common.collect.ImmutableSet;
 import uk.ac.ebi.atlas.search.SemanticQuery;
 import uk.ac.ebi.atlas.search.SemanticQueryTerm;
-import uk.ac.ebi.atlas.search.analyticsindex.AnalyticsSearchService;
+import uk.ac.ebi.atlas.solr.analytics.AnalyticsSearchService;
 import uk.ac.ebi.atlas.solr.bioentities.query.SolrQueryService;
 
 import javax.inject.Inject;
