+++ base/src/test/java/uk/ac/ebi/atlas/search/analyticsindex/FullAnalyticsSearchServiceIT.java
@@ -7,6 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import uk.ac.ebi.atlas.solr.analytics.FullAnalyticsSearchService;
 
 import javax.inject.Inject;
 import java.util.Set;
