+++ base/src/test/java/uk/ac/ebi/atlas/solr/analytics/differential/DifferentialAnalyticsFacetsReaderTest.java
