+++ base/src/test/java/uk/ac/ebi/atlas/solr/analytics/differential/DifferentialAnalyticsSearchDaoIT.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas.search.analyticsindex.differential;
+package uk.ac.ebi.atlas.solr.analytics.differential;
 
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.ReadContext;
