+++ gxa/src/main/java/uk/ac/ebi/atlas/sitemaps/SitemapController.java
@@ -8,7 +8,7 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import uk.ac.ebi.atlas.search.analyticsindex.AnalyticsSearchService;
+import uk.ac.ebi.atlas.solr.analytics.AnalyticsSearchService;
 
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletResponse;
