+++ gxa/src/test/java/uk/ac/ebi/atlas/solr/analytics/baseline/BaselineAnalyticsFacetsReaderTest.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas.search.analyticsindex.baseline;
+package uk.ac.ebi.atlas.solr.analytics.baseline;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
