+++ gxa/src/test/resources/uk/ac/ebi/atlas/solr/analytics/baseline/baseline.heatmap.pivot.response2.json
